+++ app/src/js/components/Forms/form.js
@@ -5,7 +5,8 @@ import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
 import {
   interval,
-  getForm
+  getForm,
+  getRequest
 } from '../../actions';
 import { get } from 'object-path';
 import {
@@ -66,9 +67,13 @@ class FormOverview extends React.Component {
   reload (immediate, timeout) {
     timeout = timeout || updateInterval;
     const formId = this.props.match.params.formId;
+    const requestId = this.props.location.search.split('=')[1];
     const { dispatch } = this.props;
     if (this.cancelInterval) { this.cancelInterval(); }
-    this.cancelInterval = interval(() => dispatch(getForm(formId)), timeout, immediate);
+    this.cancelInterval = interval(() => {
+      dispatch(getRequest(requestId));
+      dispatch(getForm(formId));
+    }, timeout, immediate);
   }
 
   navigateBack () {
@@ -88,11 +93,33 @@ class FormOverview extends React.Component {
   }
 
   getAnswer (id) {
-    const formId = this.props.match.params.formId;
-    const record = this.props.forms.map[formId];
-    const form = record.data;
-    if (typeof form.form_data !== 'undefined') {
-      return (form.form_data[id]);
+    if (typeof this.props.requests.detail.data.form_data[id] !== 'undefined' && this.props.requests.detail.data.form_data[id] !== '') {
+      return (this.props.requests.detail.data.form_data[id]);
+    } else {
+      return 'no answer';
+    }
+  }
+
+  getBbox (id) {
+    let out = '';
+    const n = `${id}_north`;
+    const s = `${id}_south`;
+    const e = `${id}_east`;
+    const w = `${id}_west`;
+    if (typeof this.props.requests.detail.data.form_data[n] !== 'undefined' && this.props.requests.detail.data.form_data[n] !== '') {
+      out += `N:  ${this.props.requests.detail.data.form_data[n]} `;
+    }
+    if (typeof this.props.requests.detail.data.form_data[e] !== 'undefined' && this.props.requests.detail.data.form_data[e] !== '') {
+      out += `E:  ${this.props.requests.detail.data.form_data[e]} `;
+    }
+    if (typeof this.props.requests.detail.data.form_data[s] !== 'undefined' && this.props.requests.detail.data.form_data[s] !== '') {
+      out += `S:  ${this.props.requests.detail.data.form_data[s]} `;
+    }
+    if (typeof this.props.requests.detail.data.form_data[w] !== 'undefined' && this.props.requests.detail.data.form_data[w] !== '') {
+      out += `W:  ${this.props.requests.detail.data.form_data[w]}`;
+    }
+    if (out !== '') {
+      return out;
     } else {
       return 'no answer';
     }
@@ -119,12 +146,20 @@ class FormOverview extends React.Component {
               sectionQuestions.push(<li key={this.getRandom()}>{question[b].text}</li>);
               if (question[b].inputs) {
                 for (const a in question[b].inputs) {
-                  sectionQuestions.push(
-                    <li key={ this.getRandom() } style={{ marginTop: '3px', marginBottom: '3px' }}>
-                      <div key={this.getRandom()} style={{ width: '22.5%', display: 'inline-block', float: 'left' }}>{question[b].inputs[a].label}:</div>
-                      <div key={this.getRandom()}>{this.getAnswer(question[b].inputs[a].control_id)}</div>
-                    </li>
-                  );
+                  if (question[b].inputs[a].type === 'bbox') {
+                    sectionQuestions.push(
+                      <li key={ this.getRandom() } style={{ marginTop: '3px', marginBottom: '3px' }}>
+                        <div key={this.getRandom()}>{this.getBbox(question[b].inputs[a].control_id)}</div>
+                      </li>
+                    );
+                  } else {
+                    sectionQuestions.push(
+                      <li key={ this.getRandom() } style={{ marginTop: '3px', marginBottom: '3px' }}>
+                        <div key={this.getRandom()} style={{ width: '22.5%', display: 'inline-block', float: 'left' }}>{question[b].inputs[a].label}:</div>
+                        <div key={this.getRandom()}>{this.getAnswer(question[b].inputs[a].control_id)}</div>
+                      </li>
+                    );
+                  }
                 }
               }
             }
@@ -145,7 +180,18 @@ class FormOverview extends React.Component {
   render () {
     const formId = this.props.match.params.formId;
     const record = this.props.forms.map[formId];
-
+    const requestId = this.props.location.search.split('=')[1];
+    let daacId = '';
+    if (typeof this.props.requests.detail.data !== 'undefined') {
+      daacId = this.props.requests.detail.data.daac_id;
+    }
+    let thisFormUrl = `${_config.formsUrl}?formId=${formId}`;
+    if (requestId !== '') {
+      thisFormUrl += `&requestId=${requestId}`;
+    }
+    if (daacId !== '') {
+      thisFormUrl += `&group=${daacId}`;
+    }
     if (!record || (record.inflight && !record.data)) {
       return <Loading />;
     } else if (record.error) {
@@ -168,8 +214,7 @@ class FormOverview extends React.Component {
       <div className='page__component'>
         <section className='page__section page__section__header-wrapper'>
           <h1 className='heading--large heading--shared-content with-description'>{form.long_name} (Verson {form.version})</h1>
-          {lastUpdated(form.created_at)}
-          <a className='button button--small button--green button--edit form-group__element--right' href={_config.formsUrl}>Edit</a>
+          <a className='button button--small button--green button--edit form-group__element--right' href={thisFormUrl}>Edit</a>
         </section>
 
         <section className='page__section'>
@@ -200,13 +245,17 @@ FormOverview.propTypes = {
   match: PropTypes.object,
   dispatch: PropTypes.func,
   forms: PropTypes.object,
+  requests: PropTypes.object,
   logs: PropTypes.object,
-  history: PropTypes.object
+  history: PropTypes.object,
+  location: PropTypes.object,
+  params: PropTypes.object
 };
 
 FormOverview.displayName = 'FormElem';
 
 export default withRouter(connect(state => ({
   forms: state.forms,
+  requests: state.requests,
   logs: state.logs
 }))(FormOverview));