+++ app/src/js/components/Requests/request.js
@@ -4,9 +4,9 @@ import PropTypes from 'prop-types';
 import { Link, withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
 import {
-  getSubmission,
-  deleteSubmission,
-  applyWorkflowToSubmission,
+  getRequest,
+  // deleteRequest,
+  applyWorkflowToRequest,
   listWorkflows
 } from '../../actions';
 import {
@@ -41,6 +41,14 @@ const metaAccessors = [
     property: 'step_name'
   },
   {
+    label: 'Data Publication Request',
+    property: 'data_publication_request'
+  },
+  {
+    label: 'Data Production Information',
+    property: 'data_production_information'
+  },
+  {
     label: 'Created',
     accessor: d => {
       return (shortDateNoTimeYearFirst(d));
@@ -66,7 +74,7 @@ const metaAccessors = [
   }
 ];
 
-class SubmissionOverview extends React.Component {
+class RequestOverview extends React.Component {
   constructor () {
     super();
     this.reload = this.reload.bind(this);
@@ -84,20 +92,20 @@ class SubmissionOverview extends React.Component {
   }
 
   componentDidMount () {
-    const { submissionId } = this.props.match.params;
+    const { requestId } = this.props.match.params;
     const { dispatch } = this.props;
     // This causes a repeating query for workflows cluttering up the logs.
     // Commenting out until we add applyWorkflow capability
     // this.cancelInterval = interval(this.queryWorkflows, updateInterval, true);
-    dispatch(getSubmission(submissionId));
+    dispatch(getRequest(requestId));
   }
 
   reload (immediate, timeout) {
     // timeout = timeout || updateInterval;
-    // const submissionId = this.props.match.params.submissionId;
+    // const requestId = this.props.match.params.requestId;
     // const { dispatch } = this.props;
     // if (this.cancelInterval) { this.cancelInterval(); }
-    // this.cancelInterval = interval(() => dispatch(getSubmission(submissionId)), timeout, immediate);
+    // this.cancelInterval = interval(() => dispatch(getRequest(requestId)), timeout, immediate);
   }
 
   fastReload () {
@@ -115,28 +123,28 @@ class SubmissionOverview extends React.Component {
   }
 
   applyWorkflow () {
-    const { submissionId } = this.props.match.params;
+    const { requestId } = this.props.match.params;
     const { workflow } = this.state;
-    this.props.dispatch(applyWorkflowToSubmission(submissionId, workflow));
+    this.props.dispatch(applyWorkflowToRequest(requestId, workflow));
   }
 
   delete () {
-    const { submissionId } = this.props.match.params;
-    this.props.dispatch(deleteSubmission(submissionId));
+    // const { requestId } = this.props.match.params;
+    // this.props.dispatch(deleteRequest(requestId));
   }
 
-  // This method is unnecessary now, it checks for any errors on any of the submissions queried so far,
-  // since this is a detailed view of a single submission we are only concerned with an error for that one
+  // This method is unnecessary now, it checks for any errors on any of the requests queried so far,
+  // since this is a detailed view of a single request we are only concerned with an error for that one
   // so no need to relegate the error check to a separate function
   // errors () {
-  //   const submissionId = this.props.match.params.submissionId;
+  //   const requestId = this.props.match.params.requestId;
   //   return [
-  //     get(this.props.requests.map, [submissionId, 'error']),
-  //     get(this.props.requests.reprocessed, [submissionId, 'error']),
-  //     get(this.props.requests.reingested, [submissionId, 'error']),
-  //     get(this.props.requests.executed, [submissionId, 'error']),
-  //     get(this.props.requests.removed, [submissionId, 'error']),
-  //     get(this.props.requests.deleted, [submissionId, 'error'])
+  //     get(this.props.requests.map, [requestId, 'error']),
+  //     get(this.props.requests.reprocessed, [requestId, 'error']),
+  //     get(this.props.requests.reingested, [requestId, 'error']),
+  //     get(this.props.requests.executed, [requestId, 'error']),
+  //     get(this.props.requests.removed, [requestId, 'error']),
+  //     get(this.props.requests.deleted, [requestId, 'error'])
   //   ].filter(Boolean);
   // }
 
@@ -156,7 +164,7 @@ class SubmissionOverview extends React.Component {
   }
 
   render () {
-    const { submissionId } = this.props.match.params;
+    const { requestId } = this.props.match.params;
     const record = this.props.requests.detail;
     const request = record.data || false;
 
@@ -164,10 +172,10 @@ class SubmissionOverview extends React.Component {
       text: 'Delete',
       action: this.delete,
       disabled: !!request.submitted,
-      status: get(this.props.requests.deleted, [submissionId, 'status']),
+      status: get(this.props.requests.deleted, [requestId, 'status']),
       success: this.navigateBack,
       confirmAction: true,
-      confirmText: deleteText(submissionId)
+      confirmText: deleteText(requestId)
     }]; */
 
     const breadcrumbConfig = [
@@ -180,7 +188,7 @@ class SubmissionOverview extends React.Component {
         href: '/requests'
       },
       {
-        label: submissionId,
+        label: requestId,
         active: true
       }
     ];
@@ -192,7 +200,7 @@ class SubmissionOverview extends React.Component {
         </section>
 
         <section className='page__section page__section__header-wrapper'>
-          <h1 className='heading--large heading--shared-content with-description width--three-quarters'>{submissionId}</h1>
+          <h1 className='heading--large heading--shared-content with-description width--three-quarters'>{requestId}</h1>
           {/* <AsyncCommands config={dropdownConfig} /> */}
           { request && lastUpdated(request.last_change, 'Updated') }
 
@@ -204,7 +212,7 @@ class SubmissionOverview extends React.Component {
 
         <section className='page__section'>
           <div className='heading__wrapper--border'>
-            <h2 className='heading--medium with-description'>{strings.submission_overview}</h2>
+            <h2 className='heading--medium with-description'>{strings.request_overview}</h2>
           </div>
           { record.inflight ? <Loading />
             : record.error ? <ErrorReport report={record.error} />
@@ -225,10 +233,10 @@ class SubmissionOverview extends React.Component {
 
         <section className='page__section'>
           <LogViewer
-            query={{ q: submissionId }}
+            query={{ q: requestId }}
             dispatch={this.props.dispatch}
             logs={this.props.logs}
-            notFound={`No recent logs for ${submissionId}`}
+            notFound={`No recent logs for ${requestId}`}
           />
         </section> */}
       </div>
@@ -236,7 +244,7 @@ class SubmissionOverview extends React.Component {
   }
 }
 
-SubmissionOverview.propTypes = {
+RequestOverview.propTypes = {
   match: PropTypes.object,
   dispatch: PropTypes.func,
   requests: PropTypes.object,
@@ -246,14 +254,14 @@ SubmissionOverview.propTypes = {
   workflowOptions: PropTypes.array
 };
 
-SubmissionOverview.defaultProps = {
+RequestOverview.defaultProps = {
   skipReloadOnMount: false
 };
 
-export { SubmissionOverview };
+export { RequestOverview };
 
 export default withRouter(connect(state => ({
   requests: state.requests,
   workflowOptions: workflowOptionNames(state),
   logs: state.logs
-}))(SubmissionOverview));
+}))(RequestOverview));