+++ app/src/js/components/Submissions_OLD/list.js
@@ -8,7 +8,7 @@ import {
   // clearSubmissionsSearch,
   // filterSubmissions,
   // clearSubmissionsFilter,
-  listSubmissions,
+  listRequests,
   // getOptionsSubmissionName,
   listWorkflows,
   applyWorkflowToSubmission,
@@ -52,7 +52,7 @@ class AllSubmissions extends React.Component {
 
   componentDidMount () {
     this.cancelInterval = interval(this.queryWorkflows, updateInterval, true);
-    this.props.dispatch(listSubmissions);
+    this.props.dispatch(listRequests);
   }
 
   componentWillUnmount () {
@@ -87,8 +87,8 @@ class AllSubmissions extends React.Component {
     this.setState({ workflow });
   }
 
-  applyWorkflow (submissionId) {
-    return applyWorkflowToSubmission(submissionId, this.state.workflow);
+  applyWorkflow (requestId) {
+    return applyWorkflowToSubmission(requestId, this.state.workflow);
   }
 
   getExecuteOptions () {
@@ -119,7 +119,7 @@ class AllSubmissions extends React.Component {
     const view = this.getView();
     const displayCaseView = displayCase(view);
     const statusOpts = (view === 'all') ? statusOptions : null;
-    const tableSortIdx = view === 'failed' ? 'submissionId' : 'timestamp';
+    const tableSortIdx = view === 'failed' ? 'requestId' : 'timestamp';
     const breadcrumbConfig = [
       {
         label: 'Dashboard Home',
@@ -150,7 +150,7 @@ class AllSubmissions extends React.Component {
         <section className='page__section'>
           <List
             list={list}
-            action={listSubmissions}
+            action={listRequests}
             tableColumns={view === 'failed' ? errorTableColumns : tableColumns}
             query={query}
             rowId='id'
@@ -222,7 +222,7 @@ AllSubmissions.propTypes = {
   onQueryChange: PropTypes.func
 };
 
-export { listSubmissions };
+export { listRequests };
 
 export default withRouter(connect(state => ({
   logs: state.logs,