+++ app/src/js/components/Submissions_OLD/overview.js
@@ -10,7 +10,7 @@ import {
   // clearSubmissionsSearch,
   // filterSubmissions,
   // clearSubmissionsFilter,
-  listSubmissions,
+  listRequests,
   // filterStages,
   // filterStatuses,
   // clearStagesFilter,
@@ -70,7 +70,7 @@ class SubmissionsOverview extends React.Component {
   componentDidMount () {
     this.cancelInterval = interval(this.queryMeta, updateInterval, true);
     const { dispatch } = this.props;
-    dispatch(listSubmissions);
+    dispatch(listRequests);
   }
 
   componentWillUnmount () {
@@ -112,8 +112,8 @@ class SubmissionsOverview extends React.Component {
     this.setState({ workflow });
   }
 
-  applyWorkflow (submissionId) {
-    return applyWorkflowToSubmission(submissionId, this.state.workflow);
+  applyWorkflow (requestId) {
+    return applyWorkflowToSubmission(requestId, this.state.workflow);
   }
 
   getExecuteOptions () {
@@ -157,7 +157,7 @@ class SubmissionsOverview extends React.Component {
           </div>
           <List
             list={list}
-            action={listSubmissions}
+            action={listRequests}
             tableColumns={tableColumns}
             query={this.generateQuery()}
             rowId='id'
@@ -169,7 +169,7 @@ class SubmissionsOverview extends React.Component {
                 options={get(dropdowns, ['name', 'options'])}
                 action={filterSubmissions}
                 clear={clearSubmissionsFilter}
-                paramKey='submissionId'
+                paramKey='requestId'
                 label={strings.request}
                 inputProps={{
                   placeholder: 'All',