+++ app/src/js/config/index.js
@@ -11,8 +11,6 @@ const baseConfig = {
   requireEarthdataLogin: false,
   minCompatibleApiVersion: '1.0.0',
   oauthMethod: 'earthdata',
-  formsUrl: 'http://localhost:8081/interest/',
-  overviewUrl: 'https://pub.earthdata.nasa.gov/',
   basepath: '/',
   graphicsPath: '/src/assets/images/',